requirements.txt
python-pptx
sqlite3
nltk
spacy